requirements.txt
PyYAML==6.0.1
black==23.7.0
